// src/app/shared/enums/dialog-messages-enum.ts
export enum DialogMessageEnum{

    CONFIRM_TITLE = 'Bestätigen Sie den Vorgang',
    CONFIRM_MESSAGE = 'Möchten Sie diese Aktion durchführen?',
    DELETE_USER_CONFIRM =  'Möchten Sie diesen Benutzer löschen?',
    DELETE_CATEGORY_CONFIRM = 'Möchten Sie diese Kategrie löschen?',
    DELETE_DEVICE_CONFIRM = 'Möchten Sie dieses Gerät löschen?',
    DELETE_PHONE_CONFIRM = 'Möchten Sie dieses Handy löschen?',
    RESTORE_DEVICE_CONFIRM = 'Möchten Sie dieses Gerät wiederherstellen?',
    RESTORE_PHONE_CONFIRM = 'Möchten Sie dieses Handy wiederherstellen?',
    LOGIN_ERROR = 'Anmeldeversuch war nicht erfolgreich. Überprüfen Sie ihre Daten',
    LOGIN_SUCCESS = 'Erfolgreich angemeldet!',
    CREATE_USER_SUCCESS = 'Benutzer wurde erfolgreich angelegt!',
    CREATE_USER_ERROR = 'Benutzer konnte nicht angelegt werden!',
    DELETE_USER_ERROR = 'Benutzer konnte nicht gelöscht werden',
    DELETE_USER_SUCCESS = 'Benutzer wurde gelöscht.',
    DELETE_MESSAGE_CONFIRM = 'Möchten Sie den Eintrag löschen?',
    SAVE_SUCCESS = 'Eintrag wurde gespeichert',
    SAVE_ERROR = 'Eintrag konnte nicht gespeichert werden!',
    DELETE_SUCCESS = 'Eintrag wurde gelöscht',
    DELETE_ERROR = 'Eintrag konnte nicht gelöscht werden',
    RESTORE_SUCCESS = 'Wiederherstellung war erfolgreich',
    RESTORE_ERROR = 'Wiederherstellung war nicht erfolgreich',
    ASSIGNTO_SUCCESS = 'Übernahme war erfolgreich!',
    ASSIGNTO_ERROR = 'Abgabe war nicht möglich!',
    REVOKE_SUCCESS = 'Abgabe war erfolgreich!',
    REVOKE_ERROR = 'Abgabe war nicht erfolgreich!',
}